Add server auth redirect logic
